+++ public/sw.js
@@ -1,0 +1,1 @@
+if(!self.define){let e,s={};const c=(c,a)=>(c=new URL(c+".js",a).href,s[c]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()})).then((()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e})));self.define=(a,n)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let t={};const r=e=>c(e,i),d={module:{uri:i},exports:t,require:r};s[i]=Promise.all(a.map((e=>d[e]||r(e)))).then((e=>(n(...e),t)))}}define(["./workbox-4754cb34"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/Google.png",revision:"ad8ea80a91d4122425fd7a457a083e20"},{url:"/Instagram.png",revision:"2876e7e43264b38b52a7ababa6d068ed"},{url:"/Navbar_Title.png",revision:"f31203959b8b1823deeaf55d04177d9b"},{url:"/PRODUCTIMAGE.png",revision:"0df8da74308dcab14e015f8b335537c7"},{url:"/PRODUCTIMAGE2x.png",revision:"4d5426f594b2b66f37f3992085870b68"},{url:"/TikTok.png",revision:"20dd3d20f24a7680120d4e2142f15682"},{url:"/_next/static/ZgC4nJyXmf5UJDcTmrJVL/_buildManifest.js",revision:"edd4dd131400c4e0f14576f27cc1f85c"},{url:"/_next/static/ZgC4nJyXmf5UJDcTmrJVL/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/019d026b-0a07340a5b0cf043.js",revision:"0a07340a5b0cf043"},{url:"/_next/static/chunks/2e3a845b-fe8e1f07a79e60b2.js",revision:"fe8e1f07a79e60b2"},{url:"/_next/static/chunks/36d2f571-62ea8ca92c877265.js",revision:"62ea8ca92c877265"},{url:"/_next/static/chunks/41155975-ae9802db4350e28e.js",revision:"ae9802db4350e28e"},{url:"/_next/static/chunks/4577d2ec-fd82d99404cca833.js",revision:"fd82d99404cca833"},{url:"/_next/static/chunks/5d416436-a2df601d02b99c14.js",revision:"a2df601d02b99c14"},{url:"/_next/static/chunks/664-fe029c8b38e64ec1.js",revision:"fe029c8b38e64ec1"},{url:"/_next/static/chunks/675-f6b4bb7aa5cb9513.js",revision:"f6b4bb7aa5cb9513"},{url:"/_next/static/chunks/68c0a17d-211a1d11c6ab7ca5.js",revision:"211a1d11c6ab7ca5"},{url:"/_next/static/chunks/6b2193de-3b42226129822559.js",revision:"3b42226129822559"},{url:"/_next/static/chunks/737a5600-a1d7a7e8255c16ff.js",revision:"a1d7a7e8255c16ff"},{url:"/_next/static/chunks/cb355538-cbb9ca37dd0ccc5d.js",revision:"cbb9ca37dd0ccc5d"},{url:"/_next/static/chunks/fea29d9f-d9b368cbd5826ef3.js",revision:"d9b368cbd5826ef3"},{url:"/_next/static/chunks/framework-64ad27b21261a9ce.js",revision:"64ad27b21261a9ce"},{url:"/_next/static/chunks/main-06e3bd28c8e4f891.js",revision:"06e3bd28c8e4f891"},{url:"/_next/static/chunks/pages/_app-552b582cd978b364.js",revision:"552b582cd978b364"},{url:"/_next/static/chunks/pages/_error-77823ddac6993d35.js",revision:"77823ddac6993d35"},{url:"/_next/static/chunks/pages/index-6739a99835d66844.js",revision:"6739a99835d66844"},{url:"/_next/static/chunks/pages/login-2de228d0b778d851.js",revision:"2de228d0b778d851"},{url:"/_next/static/chunks/pages/preview-product-f70df185bdd6003c.js",revision:"f70df185bdd6003c"},{url:"/_next/static/chunks/pages/product-page-6f88633a501290c2.js",revision:"6f88633a501290c2"},{url:"/_next/static/chunks/pages/ssg/onboarding-4b89ffb6f26c1c1b.js",revision:"4b89ffb6f26c1c1b"},{url:"/_next/static/chunks/pages/ssg/welcome-d66a5cc23ecbb366.js",revision:"d66a5cc23ecbb366"},{url:"/_next/static/chunks/pages/ssr/onboarding-d01ab79d09440de8.js",revision:"d01ab79d09440de8"},{url:"/_next/static/chunks/pages/ssr/welcome-e9fbfe67f9f84c05.js",revision:"e9fbfe67f9f84c05"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-8fa1640cc84ba8fe.js",revision:"8fa1640cc84ba8fe"},{url:"/_next/static/css/27b4d6b789fddb24.css",revision:"27b4d6b789fddb24"},{url:"/_next/static/css/66a07cf34e94cf31.css",revision:"66a07cf34e94cf31"},{url:"/_next/static/media/4473ecc91f70f139-s.p.woff",revision:"78e6fc13ea317b55ab0bd6dc4849c110"},{url:"/_next/static/media/463dafcda517f24f-s.p.woff",revision:"cbeb6d2d96eaa268b4b5beb0b46d9632"},{url:"/add_a_photo.png",revision:"41782a05c4061e2c418b33bbb3ed70f9"},{url:"/add_image.png",revision:"c1bc62bfa679971dbc18988f029a68ff"},{url:"/add_photo_alternate.png",revision:"3a5291e0608c43e1b3e368c5940345ec"},{url:"/check_circle.svg",revision:"3929938779341578e1bbc2929e5ded72"},{url:"/favicon.ico",revision:"c30c7d42707a47a3f4591831641e50dc"},{url:"/gucci.png",revision:"8870f88a50a9ece783f145073bc40d60"},{url:"/l-nav.png",revision:"bc72416afc89a4ada2cbe5ca3341796c"},{url:"/l-nav2.png",revision:"248a712540f5c5540979d83b85060b55"},{url:"/sale.svg",revision:"8153d26e010ad9cc8e6fdd14268e38f8"},{url:"/social1.png",revision:"15768c8a2cc771c334b2da069af0af71"},{url:"/social2.png",revision:"f4bc259ee24401cb71be6ed3f61f2641"},{url:"/social3.png",revision:"c18617cd1cba60380288f43c19684a0d"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:a})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
